Remove elevation from tomography, adjust priors
In compound_poisson.downscale.Downscale, the member variable square_error does not use the elelevation.
The prior for the GP has been parameterised to be renamed to gp_variance from gp_precision. The prior for gp_variance is an exp, defined in compound_poisson.mcmc.target.
The prior simulators has been updated for the new prior.